began restructuring frontend, seperated styled components
