+++ frontend/src/App_styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export const AppDiv = styled.div`
+   background: green;
+
+`
