+++ frontend/src/index.js
@@ -2,15 +2,15 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
 
-import "./index.scss";
+import "./reset.css";
 import App from "./App";
-import Auth from './components/Auth/Auth1'
+// import Auth from './components/Auth/Auth1'
 
-const auth = new Auth()
+// const auth = new Auth()
 
 ReactDOM.render(
   <Router>
-    <App 
+    <App
       // auth={auth}
      />
   </Router>,
