+++ frontend/src/reset.css
@@ -1,3 +1,5 @@
+*/
+
 /*
 http://meyerweb.com/eric/tools/css/reset/
 v2.0 | 20110126
@@ -17,37 +19,37 @@ article, aside, canvas, details, embed,
 figure, figcaption, footer, header, hgroup,
 menu, nav, output, ruby, section, summary,
 time, mark, audio, video {
-margin: 0;
-padding: 0;
-border: 0;
-font-size: 100%;
-font: inherit;
-vertical-align: baseline;
+    margin: 0;
+    padding: 0;
+    border: 0;
+    font-size: 100%;
+    font: inherit;
+    vertical-align: baseline;
 }
 /*
 HTML5 display-role reset for older browsers
 */
 article, aside, details, figcaption, figure,
 footer, header, hgroup, menu, nav, section {
-display: block;
+    display: block;
 }
 body {
-line-height: 1;
+    line-height: 1;
 }
 ol, ul {
-list-style: none;
+    list-style: none;
 }
 blockquote, q {
-quotes: none;
+    quotes: none;
 }
 blockquote:before, blockquote:after,
 q:before, q:after {
-content: '';
-content: none;
+    content: '';
+    content: none;
 }
 table {
-border-collapse: collapse;
-border-spacing: 0;
+    border-collapse: collapse;
+    border-spacing: 0;
 }
 
 /*
@@ -57,17 +59,16 @@ General Styles
 */
 
 html {
-  box-sizing: border-box;
-  font-family: sans-serif;
-}
-
-*, *:before, *:after {
-  box-sizing: inherit;
-  position: relative;
-  transition: all 0.2s ease;
-}
-
-html, body, #root {
-  width: 100%;
-  height: 100%;
-}
+    box-sizing: border-box;
+    font-family: sans-serif;
+    font-size: 62.5%;
+  }
+  *, *:before, *:after {
+    box-sizing: inherit;
+    position: relative;
+    transition: all 0.2s ease;
+  }
+  html, body, #root {
+    width: 100%;
+    height: 100%;
+  }
